requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
click==8.1.7
colorama==0.4.6
easyocr==1.7.2
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.1.0
networkx==3.2.1
ninja==1.11.1.1
numpy==2.0.2
openalpr==1.0
opencv-python-headless==4.10.0.84
packaging==24.2
passlib==1.7.4
pillow==11.0.0
pyasn1==0.6.1
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic_core==2.23.4
python-bidi==0.6.3
python-jose==3.3.0
python-multipart==0.0.17
PyYAML==6.0.2
rsa==4.9
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tifffile==2024.8.30
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
uvicorn==0.32.0
